chore: updating api-reports
