+++ plugins/scaffolder-react/src/routes.ts
@@ -47,11 +47,6 @@ export const legacySelectedTemplateRouteRef = createSubRouteRef({
   path: '/templates/:templateName',
 });
 
-/** @alpha */
-export const nextRouteRef = createRouteRef({
-  id: 'scaffolder/next',
-});
-
 /** @public */
 export const selectedTemplateRouteRef = createSubRouteRef({
   id: 'scaffolder/selected-template',
@@ -59,24 +54,10 @@ export const selectedTemplateRouteRef = createSubRouteRef({
   path: '/templates/:namespace/:templateName',
 });
 
-/** @alpha */
-export const nextSelectedTemplateRouteRef = createSubRouteRef({
-  id: 'scaffolder/next/selected-template',
-  parent: nextRouteRef,
-  path: '/templates/:namespace/:templateName',
-});
-
 /** @public */
 export const scaffolderTaskRouteRef = createSubRouteRef({
   id: 'scaffolder/task',
   parent: rootRouteRef,
-  path: '/tasks/:taskId',
-});
-
-/** @alpha */
-export const nextScaffolderTaskRouteRef = createSubRouteRef({
-  id: 'scaffolder/next/task',
-  parent: nextRouteRef,
   path: '/tasks/:taskId',
 });
